move get_cov_radii, indent else after return, make tests faster
